requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
aiomysql
aiosqlite
pydantic>=2.0
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator
python-multipart
alembic==1.13.1
httpx
requests
celery[redis]
